Fix incorrect types in tests
