+++ test/unit/core/sagas/products.ts
@@ -1,12 +1,13 @@
 import * as effects from 'redux-saga/effects';
 import { ActionCreators } from 'redux-undo';
 import Actions from '../../../../src/core/actions';
+import RecommendationsAdapter from '../../../../src/core/adapters/recommendations';
 import * as Events from '../../../../src/core/events';
 import Requests from '../../../../src/core/requests';
 import { Tasks as productDetailsTasks } from '../../../../src/core/sagas/product-details';
 import sagaCreator, { Tasks } from '../../../../src/core/sagas/products';
 import Selectors from '../../../../src/core/selectors';
-import { Routes } from '../../../../src/core/utils';
+import * as utils from '../../../../src/core/utils';
 import suite from '../../_suite';
 
 suite('products saga', ({ expect, spy, stub }) => {
@@ -50,7 +51,7 @@ suite('products saga', ({ expect, spy, stub }) => {
         expect(receiveProducts).to.be.calledWithExactly(response);
 
         task.next();
-        expect(saveState).to.be.calledWith(Routes.SEARCH);
+        expect(saveState).to.be.calledWith(utils.Routes.SEARCH);
       });
 
       it('should call receiveDetailsProduct if redirectSingleResult true and single record', () => {
@@ -163,6 +164,223 @@ suite('products saga', ({ expect, spy, stub }) => {
         task.next();
         expect(task.throw(error).value).to.eql(effects.put(receiveMoreProductsAction));
         expect(receiveMoreProducts).to.be.calledWith(error);
+        task.next();
+      });
+    });
+    describe('fetchNavigations()', () => {
+      it('should return two actions', () => {
+        const receiveRecommendationsNavigations = spy((val) => val);
+        const receiveRecommendationsRefinements = spy((val) => val);
+        const customerId = 'id';
+        const flux: any = {
+          config: {
+            customerId,
+            recommendations: {
+              iNav: {
+                navigations: {
+                  sort: true
+                },
+                refinements: {
+                  sort: true
+                }
+              }
+            },
+          },
+          actions: {
+            receiveRecommendationsNavigations,
+            receiveRecommendationsRefinements
+          }
+        };
+        const url = 'url';
+        const body = { a: 'b' };
+        const recommendations = {
+          result: [{ values: 'truthy' }, { values: false }, { values: 'literally truthy' }]
+        };
+        const returnVal: any = [{ values: 'truthy' }, { values: 'literally truthy' }];
+        const jsonResult = 'hello';
+
+        const buildUrl = stub(RecommendationsAdapter, 'buildUrl').returns(url);
+        const buildBody = stub(RecommendationsAdapter, 'buildBody').returns(body);
+        const fetch = stub(utils, 'fetch');
+        const task = Tasks.fetchNavigations(flux, <any>{ payload: {} });
+
+        expect(task.next().value).to.eql(effects.call(fetch, url, body));
+        expect(buildUrl).to.be.calledWith(customerId, 'refinements', 'Popular');
+        expect(task.next({ json: () => jsonResult}).value).to.eql( jsonResult );
+        expect(buildBody).to.be.calledWith({
+          size: 10,
+          window: 'day',
+        });
+        expect(task.next(recommendations).value).to.eql(effects.put(returnVal));
+        expect(receiveRecommendationsNavigations).to.be.calledWith(returnVal);
+        expect(task.next().value).to.eql(effects.put(returnVal));
+        expect(receiveRecommendationsRefinements).to.be.calledWith(returnVal);
+        task.next();
+      });
+      it('should call receive navigations only when refinements sort is false', () => {
+        const receiveRecommendationsNavigations = spy((val) => val);
+        const receiveRecommendationsRefinements = spy((val) => val);
+        const customerId = 'id';
+        const flux: any = {
+          config: {
+            customerId,
+            recommendations: {
+              iNav: {
+                navigations: {
+                  sort: true
+                },
+                refinements: {
+                  sort: false
+                }
+              }
+            },
+          },
+          actions: {
+            receiveRecommendationsNavigations,
+            receiveRecommendationsRefinements
+          }
+        };
+        const url = 'url';
+        const body = { a: 'b' };
+        const recommendations = {
+          result: [{ values: 'truthy' }, { values: false }, { values: 'literally truthy' }]
+        };
+        const returnVal: any = [{ values: 'truthy' }, { values: 'literally truthy' }];
+        const jsonResult = 'hello';
+
+        const buildUrl = stub(RecommendationsAdapter, 'buildUrl').returns(url);
+        const buildBody = stub(RecommendationsAdapter, 'buildBody').returns(body);
+        const fetch = stub(utils, 'fetch');
+        const task = Tasks.fetchNavigations(flux, <any>{ payload: {} });
+
+        expect(task.next().value).to.eql(effects.call(fetch, url, body));
+        expect(buildUrl).to.be.calledWith(customerId, 'refinements', 'Popular');
+        expect(task.next({ json: () => jsonResult}).value).to.eql( jsonResult );
+        expect(buildBody).to.be.calledWith({
+          size: 10,
+          window: 'day',
+        });
+        expect(task.next(recommendations).value).to.eql(effects.put(returnVal));
+        expect(receiveRecommendationsNavigations).to.be.calledWith(returnVal);
+        expect(receiveRecommendationsRefinements).to.not.be.called;
+        task.next();
+      });
+      it('should only call receive refinements action when navigations sort is off', () => {
+        const receiveRecommendationsNavigations = spy((val) => val);
+        const receiveRecommendationsRefinements = spy((val) => val);
+        const customerId = 'id';
+        const flux: any = {
+          config: {
+            customerId,
+            recommendations: {
+              iNav: {
+                navigations: {
+                  sort: false
+                },
+                refinements: {
+                  sort: true
+                }
+              }
+            },
+          },
+          actions: {
+            receiveRecommendationsNavigations,
+            receiveRecommendationsRefinements
+          }
+        };
+        const url = 'url';
+        const body = { a: 'b' };
+        const recommendations = {
+          result: [{ values: 'truthy' }, { values: false }, { values: 'literally truthy' }]
+        };
+        const returnVal: any = [{ values: 'truthy' }, { values: 'literally truthy' }];
+        const jsonResult = 'hello';
+
+        const buildUrl = stub(RecommendationsAdapter, 'buildUrl').returns(url);
+        const buildBody = stub(RecommendationsAdapter, 'buildBody').returns(body);
+        const fetch = stub(utils, 'fetch');
+        const task = Tasks.fetchNavigations(flux, <any>{ payload: {} });
+
+        expect(task.next().value).to.eql(effects.call(fetch, url, body));
+        expect(buildUrl).to.be.calledWith(customerId, 'refinements', 'Popular');
+        expect(task.next({ json: () => jsonResult}).value).to.eql( jsonResult );
+        expect(buildBody).to.be.calledWith({
+          size: 10,
+          window: 'day',
+        });
+        expect(task.next(recommendations).value).to.eql(effects.put(returnVal));
+        expect(receiveRecommendationsRefinements).to.be.calledWith(returnVal);
+        expect(receiveRecommendationsNavigations).to.not.be.called;
+        task.next();
+      });
+      it('should not call any actions when both navigations and refinements sort are off', () => {
+        const receiveRecommendationsNavigations = spy((val) => val);
+        const receiveRecommendationsRefinements = spy((val) => val);
+        const customerId = 'id';
+        const flux: any = {
+          config: {
+            customerId,
+            recommendations: {
+              iNav: {
+                navigations: {
+                  sort: false
+                },
+                refinements: {
+                  sort: false
+                }
+              }
+            },
+          },
+          actions: {
+            receiveRecommendationsNavigations,
+            receiveRecommendationsRefinements
+          }
+        };
+        const url = 'url';
+        const body = { a: 'b' };
+        const recommendations = {
+          result: [{ values: 'truthy' }, { values: false }, { values: 'literally truthy' }]
+        };
+        const returnVal: any = [{ values: 'truthy' }, { values: 'literally truthy' }];
+        const jsonResult = 'hello';
+
+        const buildUrl = stub(RecommendationsAdapter, 'buildUrl').returns(url);
+        const buildBody = stub(RecommendationsAdapter, 'buildBody').returns(body);
+        const fetch = stub(utils, 'fetch');
+        const task = Tasks.fetchNavigations(flux, <any>{ payload: {} });
+
+        task.next();
+        expect(receiveRecommendationsRefinements).to.not.be.called;
+        expect(receiveRecommendationsNavigations).to.not.be.called;
+        task.next();
+      });
+
+      it('should handle request failure', () => {
+        const error = new Error();
+        const receiveRecommendationsNavigationsAction: any = { a: 'b' };
+        const receiveRecommendationsRefinementsAction: any = { a: 'b' };
+        const receiveRecommendationsNavigations = spy(() => receiveRecommendationsNavigationsAction);
+        const receiveRecommendationsRefinements = spy(() => receiveRecommendationsRefinementsAction);
+        const flux: any = { config: {
+          recommendations: {
+            iNav: {
+              navigations: {
+                sort: true
+              },
+              refinements: {
+                sort: true
+              }
+            }
+          },
+        }, actions: {
+          receiveRecommendationsNavigations, receiveRecommendationsRefinements } };
+
+        const task = Tasks.fetchNavigations(flux, <any>{ payload: {} });
+        task.next();
+        expect(task.throw(error).value).to.eql(effects.put(receiveRecommendationsNavigationsAction));
+        expect(receiveRecommendationsNavigations).to.be.calledWithExactly(error);
+        task.next();
+        expect(receiveRecommendationsRefinements).to.be.calledWithExactly(error);
         task.next();
       });
     });
